Create product_inventory
V001 product inventory.
I am currently using a dictionary (with some pre made products) in the code to store all the information. I intend to change this to be stored in a text file that can be written to and eventually I change this to a database. The reasons for the text file is it will be a learning process for exploring a way of storing, writing and iterating data without the use of a database.